requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
redis
websockets
pydantic
python-dotenv
PyJWT>=2.0.0
